requirements.txt
streamlit>=1.32.0
mistralai>=0.0.8
python-dotenv>=1.0.0
pytest>=7.4.0
Pillow>=10.0.0
requests>=2.31.0
